pyproject.toml
[tool.poetry]
name = "lnkdnlm"
version = "0.1.0"
description = "LLM powered linkedin job search"
authors = ["Maqbool ur Rahim Khan <maqboolurrahim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.12.1"
streamlit = "^1.36.0"
streamlit-extras = "^0.4.3"
langchain = "^0.2.6"
langchain-community = "^0.2.5"
langchain-core = "^0.2.10"
langchainhub = "^0.1.20"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^7.4.3"
ruff = "^0.1.9"
click = "^8.1.7"
watchdog = "^4.0.1"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
